requirements.txt
pylint
pre-commit
django
djangorestframework
